create select2helperasset and move javascript to this asset
