+++ widgets/FormBuilder.php
@@ -10,6 +10,7 @@ use kartik\datetime\DateTimePicker;
 use maddoger\widgets\Select2;
 use maddoger\widgets\Select2BootstrapAsset;
 use netis\crud\assets\RelationsAsset;
+use netis\crud\assets\Select2HelperAsset;
 use netis\crud\crud\Action;
 use netis\crud\db\ActiveQuery;
 use netis\crud\db\ActiveRecord;
@@ -676,58 +677,7 @@ JavaScript;
      */
     public static function registerSelect($view)
     {
-        $script = <<<JavaScript
-(function (s2helper, $, undefined) {
-    "use strict";
-    s2helper.formatResult = function (result, container, query, escapeMarkup, depth) {
-        if (typeof depth == 'undefined') {
-            depth = 0;
-        }
-        var markup = [];
-        window.Select2.util.markMatch(result._label, query.term, markup, escapeMarkup);
-        return markup.join("");
-    };
-
-    s2helper.formatSelection = function (item) {
-        return item._label;
-    };
-
-    // generates query params
-    s2helper.data = function (term, page) {
-        return { search: term, page: page };
-    };
-
-    // builds query results from ajax response
-    s2helper.results = function (data, page) {
-        return { results: data.items, more: page < data._meta.pageCount };
-    };
-
-    s2helper.getParams = function (element) {
-        var primaryKey = element.data('relation-pk');
-        if (typeof primaryKey === 'undefined' || primaryKey === null) {
-            primaryKey = 'id';
-        }
-
-        var params = {search: {}};
-        params.search[primaryKey] = element.val();
-        return params;
-    };
-
-    s2helper.initSingle = function (element, callback) {
-        $.getJSON(element.data('select2').opts.ajax.url, s2helper.getParams(element), function (data) {
-            if (typeof data.items[0] != 'undefined') {
-                callback(data.items[0]);
-            }
-        });
-    };
-
-    s2helper.initMulti = function (element, callback) {
-        $.getJSON(element.data('select2').opts.ajax.url, s2helper.getParams(element), function (data) {callback(data.items);});
-    };
-}( window.s2helper = window.s2helper || {}, jQuery ));
-JavaScript;
-        $view->registerJs($script, View::POS_END, 'netis.s2helper');
-        Select2BootstrapAsset::register($view);
+        Select2HelperAsset::register($view);
     }
 
     /**
@@ -927,6 +877,8 @@ JavaScript;
         $relationName = Html::getAttributeName($relation);
         $prefixedFk = str_replace($relationName, $foreignKey, $relation);
         $value = Html::getAttributeValue($model, $isMany ? $relation : $prefixedFk);
+
+        Select2HelperAsset::register(Yii::$app->view);
         return [
             'class'         => Select2::className(),
             'model'         => $model,
@@ -1029,6 +981,7 @@ JavaScript;
         //we get prefix from $relation because it could be in format [3]relation and we need to have [3]foreign_key here
         $relationName = Html::getAttributeName($relation);
         $prefixedFk = str_replace($relation, $foreignKey, $relationName);
+        Select2HelperAsset::register(Yii::$app->view);
         return [
             'class'         => Select2::class,
             'model'         => $model,
